python/requirements.txt
# Python dependencies for Seva Connect backend services

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Image Processing for Certificates
Pillow==10.0.1
reportlab==4.0.4

# Database
PyMySQL==1.1.0
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
msal==1.24.1

# Payment Gateway
razorpay==1.3.0

# HTTP Requests
requests==2.31.0

# Date/Time handling
python-dateutil==2.8.2

# PDF generation
fpdf2==2.7.6
